+++ docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Essentials: Added an option to disable a given EXE card from the starter pool.
+* Essentials: Added a starter deck card preview.
 * Added character lookup to `IModCharacters`.
 
 ## 0.8.3